[ADD] Add Portal Page for logged In User
